+++ src/main/java/com/twitter/pers/bipartite/SALSASmallMem.java
@@ -146,7 +146,7 @@ public class SALSASmallMem extends PigGraphChiBase implements GraphChiProgram<Fl
 
         VertexAggregator.foreach(engine.numVertices(), graphName, new FloatPairConverter(), new ForeachCallback<FloatPair>() {
             @Override
-            public void callback(int vertexId, FloatPair vertexValue) {
+            public void callback(long vertexId, FloatPair vertexValue) {
                 if (vertexValue.first > 0) {
                     System.out.println(engine.getVertexIdTranslate().backward(vertexId)  + "\t" + vertexValue.first);
                 }
@@ -200,7 +200,7 @@ public class SALSASmallMem extends PigGraphChiBase implements GraphChiProgram<Fl
         // It would be better to have an iterator for the vertex data.
         VertexAggregator.foreach(engine.numVertices(), graphName, new FloatPairConverter(), new ForeachCallback<FloatPair>() {
             @Override
-            public void callback(int vertexId, FloatPair vertexValue) {
+            public void callback(long vertexId, FloatPair vertexValue) {
                 if (vertexValue.first > 0) {
                     results.add(new IdFloat(engine.getVertexIdTranslate().backward(vertexId), vertexValue.first));
                 }
@@ -216,12 +216,12 @@ public class SALSASmallMem extends PigGraphChiBase implements GraphChiProgram<Fl
         return new FastSharder<FloatPair, Float>(graphName, numShards, new VertexProcessor<FloatPair>() {
             @Override
             /* For lists (hubs), the vertex value will encode the total number of edges */
-            public FloatPair receiveVertexValue(int vertexId, String token) {
+            public FloatPair receiveVertexValue(long vertexId, String token) {
                 return new FloatPair(0.0f, Float.parseFloat(token));
             }
         }, new EdgeProcessor<Float>() {
             @Override
-            public Float receiveEdge(int from, int to, String token) {
+            public Float receiveEdge(long from, long to, String token) {
                 return Float.parseFloat(token);
             }
         }, new FloatPairConverter(), new FloatConverter());
