+++ src/main/java/edu/cmu/graphchi/preprocessing/VertexProcessor.java
@@ -5,6 +5,6 @@ package edu.cmu.graphchi.preprocessing;
  */
 public interface VertexProcessor  <ValueType> {
 
-    ValueType receiveVertexValue(int vertexId, String token);
+    ValueType receiveVertexValue(long vertexId, String token);
 
 }
