+++ src/main/java/edu/cmu/graphchi/util/IdInt.java
@@ -4,15 +4,15 @@ package edu.cmu.graphchi.util;
  * Tuple containg a vertex-id and an integer number.
  */
 public class IdInt {
-    int vertexId;
+    long vertexId;
     int value;
 
-    public IdInt(int vertexId, int value) {
+    public IdInt(long vertexId, int value) {
         this.vertexId = vertexId;
         this.value = value;
     }
 
-    public int getVertexId() {
+    public long getVertexId() {
         return vertexId;
     }
