+++ test/edu/cmu/graphchi/preprocessing/TestVertexIdTranslate.java
@@ -15,8 +15,8 @@ public class TestVertexIdTranslate {
             for(int numShards=(10000000 / verticesInShard + 1); numShards < 100; numShards += 13) {
                 VertexIdTranslate trans = new VertexIdTranslate(verticesInShard, numShards);
                 for(int j=0; j < 10000000; j+=(1+j)) {
-                     int vprime = trans.forward(j);
-                     int back = trans.backward(vprime);
+                     long vprime = trans.forward(j);
+                     long back = trans.backward(vprime);
                      assertEquals(j, back);
                 }
 
@@ -39,8 +39,8 @@ public class TestVertexIdTranslate {
         assertEquals(tr.getVertexIntervalLength(), trReconstr.getVertexIntervalLength());
 
         for(int j=0; j < 10000000; j+=(1+j)) {
-            int v1 = tr.forward(j);
-            int v2 = trReconstr.forward(j);
+            long v1 = tr.forward(j);
+            long v2 = trReconstr.forward(j);
             assertEquals(v1, v2);
         }
     }
